# ApiDoc/api_apollo-openapi.yaml
openapi: "3.0.3"
info:
  title: "api_apollo API"
  description: "api_apollo API"
  version: "1.0.0"
servers:
  - url: "https://api-apollo.onrender.com"
paths:
...
  /orders/{id}:
    put:
      summary: "PUT orders/{id}"
      operationId: "UpdataOrders"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RresponseOrders"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"

    get:
      summary: "GET orders/{id}"
      operationId: "GetOrdersId"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /orders:
    post:
      summary: "POST orders"
      operationId: "RegisterOrders"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RrequestOrders"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET orders"
      operationId: "GetOrders"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
components:
  schemas:
...
    HttpHeaders:
      type: "object"
      properties: { }
    ResponseEntity:
      type: "object"
      properties:
        headers:
          $ref: "#/components/schemas/HttpHeaders"
...
    RresponseOrders:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        equipment:
          type: "integer"
          format: "int32"
        address:
          type: "integer"
          format: "int32"
        shipping:
          type: "string"
    RrequestOrders:
      type: "object"
      properties:
        request:
          type: "string"
        requestDate:
          type: "string"
          format: "date-time"
        client:
          type: "integer"
          format: "int32"
        equipment:
          type: "integer"
          format: "int32"
        address:
          type: "integer"
          format: "int32"
        shipping:
          type: "string"